Some configfile parts tested and work.
It appears that ArrayAccess is very limited in PHP,
so that objects implementing it don't fully act like arrays
to take advantage of square bracket syntax.  Python is so much better in
this regard.

But my deepSet() method works great, and I can build further methods to
flesh out a nice interface.